Adapt to new schema

+++ dogebuild_c/c_plugin.py
@@ -2,7 +2,7 @@ from subprocess import call
 from dogebuild.plugins import DogePlugin
 import os
 import shutil
-from typing import List, Tuple
+from typing import List, Tuple, Dict
 
 from dogebuild_c.gcc_wrapper import GccWrapper, BinaryType
 
@@ -15,12 +15,12 @@ class CPlugin(DogePlugin):
 
         self.gcc = GccWrapper()
 
-        self.add_task('compile', self.compile, phase='compile')
+        self.add_task('compile', self.compile, phase='sources')
 
         self.add_task('test', self.test, phase='test')
         self.add_dependency('test', ['compile'])
 
-        self.add_task('link', self.link, phase='link')
+        self.add_task('link', self.link, phase='build')
         self.add_dependency('link', ['test'])
 
         self.add_task('run', self.run, phase='run')
@@ -40,23 +40,31 @@ class CPlugin(DogePlugin):
         self.o_files = []
         self.test_exclude = kwargs.get('test_exclude', [])
 
-    def compile(self) -> int:
+    def compile(self) -> Tuple[int, Dict[str, List[str]]]:
         code, o_files = self.gcc.compile(self.build_dir, self.type, self.src, [])
         if code:
-            return code
+            return code, {}
         else:
             self.o_files = o_files
-            return 0
+            return 0, {'o_files': o_files}
 
-    def test(self) -> int:
+    def test(self) -> Tuple[int, Dict[str, List[str]]]:
         if len(self.test_src) == 0:
-            return 0
+            return 0, {}
 
         self.gcc.copy_header(self.build_dir, self.headers)
 
-        code, test_o_files = self.gcc.compile(self.test_build_dir, BinaryType.EXECUTABLE, self.test_src, [os.path.join(self.build_dir, 'headers')])
+        dependencies_headers = []
+        for dependency in self.dependencies + self.test_dependencies:
+            dependencies_headers += map(lambda p: os.path.join(dependency.folder, p), dependency.artifacts.get('headers_directory', []))
+
+        dependencies_library = []
+        for dependency in self.dependencies + self.test_dependencies:
+            dependencies_library += map(lambda p: os.path.join(dependency.folder, p), dependency.artifacts.get('static_library', []))
+
+        code, test_o_files = self.gcc.compile(self.test_build_dir, BinaryType.EXECUTABLE, self.test_src, [os.path.join(self.build_dir, 'headers')] + dependencies_headers)
         if code:
-            return code
+            return code, {}
 
         # Fixme
         exclude_o_files = []
@@ -68,32 +76,46 @@ class CPlugin(DogePlugin):
                 file_path = os.path.join(self.build_dir, src)
                 exclude_o_files.append(file_path)
 
-        code, test_out_file = self.gcc.link(self.test_build_dir, list(set(self.o_files) - set(exclude_o_files)) + test_o_files, self.test_out, BinaryType.EXECUTABLE)
+        code, test_out_file = self.gcc.link(self.test_build_dir, BinaryType.EXECUTABLE, self.test_out, list(set(self.o_files) - set(exclude_o_files)) + test_o_files, dependencies_library)
         if code:
-            return code
+            return code, {}
 
-        return call([test_out_file])
+        return call([test_out_file]), {'test_executable': [test_out_file]}
 
-    def link(self) -> int:
-        code, out_file = self.gcc.link(self.build_dir, self.o_files, self.out, self.type)
+    def link(self) -> Tuple[int, Dict[str, List[str]]]:
+        code, out_file = self.gcc.link(self.build_dir, self.type, self.out, self.o_files, [])
         if code:
-            return code
+            return code, {}
 
         self.out_file = out_file
 
+        headers_directory = []
         if self.type in [BinaryType.STATIC_LIBRARY, BinaryType.DYNAMIC_LIBRARY]:
-            self.gcc.copy_header(self.build_dir, self.headers)
+            _, headers_directory = self.gcc.copy_header(self.build_dir, self.headers)
 
-        return 0
+        if self.type is BinaryType.EXECUTABLE:
+            artifact = {'executable': [out_file]}
+        elif self.type is BinaryType.STATIC_LIBRARY:
+            artifact = {'static_library': [out_file]}
+        elif self.type is BinaryType.DYNAMIC_LIBRARY:
+            artifact = {'dynamic_library': [out_file]}
+        else:
+            raise NotImplementedError()
 
-    def run(self) -> int:
+        return 0, dict(artifact, headers_directory=[headers_directory])
+
+    def run(self) -> Tuple[int, Dict[str, List[str]]]:
         if self.type is BinaryType.EXECUTABLE:
             command = [self.out_file]
-            return call(command)
+            return call(command), {'executable': [self.out_file]}
         else:
             print('Type {} is not executable'.format(self.type))
-            return 0
+            return 0, {}
 
-    def clean(self):
-        shutil.rmtree(self.build_dir)
-        shutil.rmtree(self.test_build_dir)
+    def clean(self) -> Tuple[int, Dict[str, List[str]]]:
+        if os.path.exists(self.build_dir):
+            shutil.rmtree(self.build_dir)
+        if os.path.exists(self.test_build_dir):
+            shutil.rmtree(self.test_build_dir)
+
+        return 0, {}
